pyproject.toml
[tool.poetry]
name = "python_template"
version = "0.1.0"
description = ""
authors = ["Jordan Hayes <jordanhayes98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
mypy = "^1.8.0"
python = "^3.13"
pytest = "^7.4.4"
ruff = "^0.1.14"
flask = "^3.1.1"
google-genai = "^1.25.0"
google = "^3.0.0"
python-dotenv = "^1.1.1"
discord-py = "^2.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
enable_incomplete_feature = ["NewGenericSyntax"]
ignore_missing_imports = true

[tool.mypy."google.*"]
ignore_errors = true

[tool.ruff]
line-length = 80
indent-width = 4
